// src/types/doc.types.ts
export interface IDocResponse {
	ndoc: string
	station_otpr: string
	station_nazn: string
	kotpr: number
	notpr: string
	kpol: number
	npol: string
	way_code: string
	dat_oform: string
	dat_ceh_otpr: string
	dat_st_otpr: string
	dat_st_nazn: string
	dat_ceh_nazn: string
	usr_ceh_otpr: string
	usr_st_otpr: string
	usr_st_nazn: string
	usr_ceh_nazn: string
	num_doc: number
	user_id: string
}